[WIP]- basic refactor
