+++ .env
@@ -1,1 +1,1 @@
-ELEPHANTSQL_URL="*************"
+ELEPHANTSQL_URL="postgres://fbkemlaw:2SnA_TUERMz8nVxsTuNLV1tlN3xEuzCV@stampy.db.elephantsql.com:5432/fbkemlaw"
