+++ main.go
@@ -9,16 +9,11 @@ import (
 	"log"
 	"net/http"
 	"os"
+	"go-book/models"
 )
 
-type Book struct {
-	ID int `json:"id"`
-	Title string `json:"title"`
-	Author string `json:"author"`
-	Year string `json:"year"`
-}
 
-var books [] Book
+var books [] models.Book
 var db *sql.DB
 func init()  {
 	gotenv.Load()
@@ -70,8 +65,8 @@ func getBooks(w http.ResponseWriter, r *http.Request)  {
 
 func getBook(w http.ResponseWriter, r *http.Request)  {
 	params:=mux.Vars(r) // we have id of books in map
-	var book Book
-	books = []Book{}
+	var book models.Book
+	books = []models.Book{}
 	rows,err:=db.Query("select * from books where id=$1", params["id"])
 	logFatal(err)
 	for rows.Next(){
@@ -82,7 +77,7 @@ func getBook(w http.ResponseWriter, r *http.Request)  {
 }
 
 func addBook(w http.ResponseWriter, r *http.Request)  {
-	var book Book
+	var book models.Book
 	var bookID int
 	json.NewDecoder(r.Body).Decode(&book)
 	err:=db.QueryRow("INSERT INTO books (title,author,year) values($1,$2,$3) RETURNING id;", book.Title, book.Author,book.Year).Scan(&bookID)
@@ -92,7 +87,7 @@ func addBook(w http.ResponseWriter, r *http.Request)  {
 }
 
 func updateBook(w http.ResponseWriter, r *http.Request)  {
-	var book Book
+	var book models.models.Book
 	json.NewDecoder(r.Body).Decode(&book)
 	result,err:=db.Exec("update books set title=$1, author=$2, year=$3 where id=$4 RETURNING id", &book.Title,&book.Author,&book.Year,&book.ID)
 	logFatal(err)
